Don't let a background task call run_background_tasks()
