+++ ports/atmel-samd/background.h
@@ -29,6 +29,7 @@
 
 #include <stdbool.h>
 
+void background_tasks_reset(void);
 void run_background_tasks(void);
 void run_background_vm_tasks(void);
 bool background_tasks_ok(void);
